# Back-end/User-API/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Fitisly",
    "description": "API for the fitlisly app",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "ec2-3-14-244-98.us-east-2.compute.amazonaws.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "definitions": {
...
    "program": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "programImage": {
          "type": "string"
        },
        "exercises": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  },
  "paths": {
...
    "/api/programs": {
      "parameters": [
        {
          "name": "Authorization",
          "in": "header",
          "description": "Authorization header",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Programs"
        ],
        "summary": "Get all the programs of the database",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/program"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Programs"
        ],
        "summary": "Create a new program in the database",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "programImage",
            "description": "the image of the program",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "name",
            "type": "string",
            "description": "the name of the program"
          },
          {
            "in": "formData",
            "name": "description",
            "type": "string",
            "description": "the description of the program"
          },
          {
            "in": "formData",
            "name": "exercises",
            "type": "array",
            "description": "The list of exercises ids of the program",
            "items": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/program"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/programs/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the programs that we want to retrieve",
          "type": "integer"
        },
        {
          "name": "Authorization",
          "in": "header",
          "description": "Authorization header",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Programs"],
        "summary": "Get the program with the given ID",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/program"
            }
          },
          "404": {
            "description": "Program not found"
          },
          "500":  {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": ["Programs"],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "programImage",
            "description": "the image of the program",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "name",
            "type": "string",
            "description": "the name of the program"
          },
          {
            "in": "formData",
            "name": "description",
            "type": "string",
            "description": "the description of the program"
          },
          {
            "in": "formData",
            "name": "exercises",
            "type": "array",
            "description": "The list of exercises ids of the program",
            "items": {
              "type": "integer"
            }
          }
        ],
        "summary": "Update the program with the given ID",
        "responses": {
          "200": {
            "description": "Program updated sucessfully",
            "schema": {
              "$ref": "#/definitions/program"
            }
          },
          "404": {
            "description": "Program not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["Programs"],
        "summary": "Delete programs with the given ID",
        "responses": {
          "200": {
            "description": "Program {id} was deleted sucessfully."
          },
          "404": {
            "description": "Program not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}